requirements.txt
pytest==7.3.1
yt-dlp==2024.3.10
#ffmpeg-python==0.4.7 
requests
faster-whisper
mcp[cli]
openai
httpx
python-dotenv 
tqdm